+++ private/language/english_utf-8.php
@@ -2016,6 +2016,7 @@ $LANG_ADMIN = array(
     'custom_autotag' => 'Custom Autotags',
     'alt_topic' => 'Alternate Topic',
     'timeout_msg' => 'Your session is about to expire for inactivity. Please select OK to refresh the session.',
+    'reset' => 'Reset',
 );
 
 ###############################################################################
@@ -2415,6 +2416,13 @@ $LANG_AM = array (
     'allowed'                   => 'Allowed',
     'namespace'                 => 'Namespace',
     'operation'                 => 'Operation',
+    'content_area'              => 'Content Area',
+    'check_all'                 => 'Check All',
+    'autotag_allowed'           => 'Auto tags allowed for ',
+    'perm_saved'                => 'Permissions have been saved',
+    'perm_instructions'         => 'Check the auto tags you will allow for the specified content area',
+    'perm_header'               => 'Auto Tag Permissions by Content Area',
+    'perm_editor'               => 'Permission Editor',
 );
 
 ###############################################################################